Expand file size from int to long to fix issue with reading files that are larger than 2.1 GB.

+++ src/main/java/com/uploadcare/api/File.java
@@ -53,7 +53,7 @@ public class File {
         return fileData.datetimeRemoved;
     }
 
-    public int getSize() {
+    public long getSize() {
         return fileData.size;
     }
